Added support for assets and storage locations
- Added ability to retrieve data from two schema
if there is an existing relationship between them. Added couple of minor
support for reports page since it is more complex than anticipated.

- Perform some code cleanup as well for easier state management. Using
tanstack query is helpful to remove some of the useEffects which do not
serve any purpose with the useHooks in place.

- Update table constants to properly fetch the storage location and applied
some tech debt fixes as well.