Updated image uploads
Fixed an issue with the image uploads where the user would need to refresh
the page to view the selected image. There was multiple api calls that
were happening which were redundant and i think the query would not
invalidate because of that.

Fixed formFactor issues for selected asset page and kept variables
consistent across all packages.

Updated package-lock.json file to support changes. The default node version
requirement in CI is 23.7.0 for node, and I suggest we keep the same for
the jsx files here as well.