Add ability to remove associations
Added ability to remove assocations between assets and categories or
maintenance plans if the selected asset is removed.

Also removes the category / maintenace plan from the favourite items list
if the category / maintenace plan is removed.

Found a bug where if a user has existing category / maintenance plan bookmarked
then the user would be able to navigate to that url and make changes that
would not be reflected in the system. Applied fix where if the selected category
or maintenace plan does not exist, the user would not be able to venture
forward rather would just see a loading indicator.