Apply fix for removing assets
A single asset could be used across many categories or many different
maintenance plans. Added fix to resolve this issue. Now assets that are
listed in various categories would also be removed since the parent asset
is no longer present.

Added navigation changes for MenuActionBar so that when a user clicks on
a selected asset / fav item, they can be re-routed to the correct a/o